Fixed: Footer , Background Gradient, Theme Adjustment
The footer is now complete. Two different background gradient is added for the canvas.
The new color is added in the theme for giving the functionlity of dark and light theme.
A small bug in the ThemeButton is fixed. The bug was due to two separate component for different viewpoint.